+++ storage_drivers/ontap/ontap_nas_qtree.go
@@ -11,6 +11,7 @@ import (
 	"sync"
 	"time"
 
+	"github.com/RoaringBitmap/roaring"
 	log "github.com/sirupsen/logrus"
 
 	trident "github.com/netapp/trident/config"
@@ -1214,6 +1215,22 @@ func (d *NASQtreeStorageDriver) getVolumeExternal(
 		Config: volumeConfig,
 		Pool:   volumeIDAttrs.ContainingAggregateName(),
 	}
+}
+
+// GetUpdateType returns a bitmap populated with updates to the driver
+func (d *NASQtreeStorageDriver) GetUpdateType(driverOrig storage.Driver) *roaring.Bitmap {
+	bitmap := roaring.New()
+	dOrig, ok := driverOrig.(*NASQtreeStorageDriver)
+	if !ok {
+		bitmap.Add(storage.InvalidUpdate)
+		return bitmap
+	}
+
+	if d.Config.DataLIF != dOrig.Config.DataLIF {
+		bitmap.Add(storage.VolumeAccessInfoChange)
+	}
+
+	return bitmap
 }
 
 type HousekeepingTask struct {