+++ storage_drivers/solidfire/solidfire_san.go
@@ -10,6 +10,7 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/RoaringBitmap/roaring"
 	log "github.com/sirupsen/logrus"
 
 	trident "github.com/netapp/trident/config"
@@ -1221,4 +1222,20 @@ func (d *SANStorageDriver) getVolumeExternal(
 		Config: volumeConfig,
 		Pool:   drivers.UnsetPool,
 	}
+}
+
+// GetUpdateType returns a bitmap populated with updates to the driver
+func (d *SANStorageDriver) GetUpdateType(driverOrig storage.Driver) *roaring.Bitmap {
+	bitmap := roaring.New()
+	dOrig, ok := driverOrig.(*SANStorageDriver)
+	if !ok {
+		bitmap.Add(storage.InvalidUpdate)
+		return bitmap
+	}
+
+	if d.Config.SVIP != dOrig.Config.SVIP {
+		bitmap.Add(storage.VolumeAccessInfoChange)
+	}
+
+	return bitmap
 }